requirements.txt
numpy==1.26.4
opencv-python==4.10.0.84
ffmpeg-python==0.2.0
Pillow==10.4.0
pyyaml==6.0.2
pydantic==2.9.2
scenedetect==0.6.4
mediapipe==0.10.14
ultralytics==8.3.0
torch==2.4.1
torchvision==0.19.1
open_clip_torch==2.26.1
google-generativeai==0.8.2
python-dotenv==1.0.1
fastapi==0.115.0
uvicorn==0.31.0
websockets==13.1
pytest==8.3.3
